+++ config/app.php
@@ -14,7 +14,7 @@ return [
     */
 
     'name' => 'Lunar Base',
-    'version' => '1.1.0',
+    'version' => '1.2.0',
     'author' => 'Cau Guanabara',
 
     /*